requirements.txt
Django>=4.2.0
Pillow>=9.5.0
django-ckeditor>=6.5.1
django-select2>=8.1.2